-- Performance improvements done by Mike Ovassianikov:  1. Replace the chunk manager's list of pending writes with a set.  This leads  to fast find on the pending writes (replace a O(n) scan with a O(logn) scan).  Under load for writes, CPU use drops by 30% leading to an equivalent perf.  boost.  2. New logic / methods in IOBuffer class to keep underlying buffers page  aligned.  3. Use scatter-gatherer io (readv / writev) with sockets in IOBuffer class  4. Implement delayed chunk meta data write, in order to reduce # of disk seeks.

+++ trunk/src/cc/qcdio/qcmutex.h
@@ -1,5 +1,5 @@
 //---------------------------------------------------------- -*- Mode: C++ -*-
-// $Id: //depot/main/platform/kosmosfs/src/cc/qcdio/qcmutex.h#1 $
+// $Id: //depot/main/platform/kosmosfs/src/cc/qcdio/qcmutex.h#2 $
 //
 // Created 2008/10/30
 // Author: Mike Ovsiannikov
@@ -28,11 +28,12 @@
 
 #include <pthread.h>
 #include <errno.h>
+#include <stdint.h>
 
 class QCMutex
 {
 public:
-    typedef long long Time;
+    typedef int64_t Time;
 
     QCMutex();
     ~QCMutex();
